--- app/data/database.py
@@ -1,14 +1,0 @@
-import aiomysql
-from app.config.settings import settings
-
-
-async def get_connection():
-    return await aiomysql.connect(
-        host=settings.DB_HOST,
-        port=settings.DB_PORT,
-        user=settings.DB_USER,
-        password=settings.DB_PASSWORD,
-        db=settings.DB_NAME,
-        charset="utf8mb4",
-        autocommit=True
-    )
