+++ python/config/setting.py
@@ -4,7 +4,7 @@ from dotenv import load_dotenv
 load_dotenv()
 
 
-class Settings:
+class Setting:
     DB_HOST = os.getenv("DB_HOST")
     DB_PORT = int(os.getenv("DB_PORT"))
     DB_USER = os.getenv("DB_USER")
@@ -13,4 +13,4 @@ class Settings:
     CHROMA_DIR = os.path.abspath("../storage/chroma_db")
 
 
-settings = Settings()
+setting = Setting()
