+++ python/domain/model/query_validity.py
@@ -1,0 +1,6 @@
+from pydantic import BaseModel
+
+# 질문 품질 필터링
+class QueryValidity(BaseModel):
+    is_valid: bool
+    reason: str
