+++ python/infrastructure/embedding_model.py
